requirements.txt
Flask==2.3.3
requests==2.31.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
gunicorn==21.2.0
python-dotenv==1.0.1
